# specs/001-gs-control-center/contracts/api.yaml
openapi: 3.1.0
info:
  title: GS Control Center API
  description: Centralized Advertising Block Management System for UCS Guest Screen
  version: 1.0.0
servers:
  - url: /api/v1
    description: Central Server API (Host 10.0.0.111)
paths:
...
  /advertising-blocks:
    get:
      summary: List advertising blocks (filterable by area and display mode)
      parameters:
        - name: area
          in: query
          schema: { type: string, enum: [FULL_SCREEN, MODE32_PROMO] }
        - name: display_mode
          in: query
          schema: { type: string, enum: [STATIC, SLIDESHOW, VIDEO] }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AdvertisingBlock'
    post:
      summary: Create new advertising block with playlist items
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, area, display_mode, playlist]
              properties:
                name: { type: string }
                description: { type: string }
                area: { type: string, enum: [FULL_SCREEN, MODE32_PROMO] }
                display_mode: { type: string, enum: [STATIC, SLIDESHOW, VIDEO] }
                playlist:
                  type: array
                  items:
                    type: object
                    required: [media_asset_id, order_index]
                    properties:
                      media_asset_id: { type: string, format: uuid }
                      order_index: { type: integer }
                      duration_seconds: { type: integer, default: 7 }
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdvertisingBlock'
  /advertising-blocks/{id}:
    get:
      summary: Get advertising block details and playlist
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdvertisingBlock'
    put:
      summary: Update advertising block, reorder playlist, or modify durations (OCC protected)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [version]
              properties:
                name: { type: string }
                description: { type: string }
                is_active: { type: boolean }
                version: { type: integer, description: Monotonically increasing OCC version counter }
                playlist:
                  type: array
                  items:
                    type: object
                    required: [media_asset_id, order_index]
                    properties:
                      media_asset_id: { type: string, format: uuid }
                      order_index: { type: integer }
                      duration_seconds: { type: integer }
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdvertisingBlock'
        '409':
          description: Optimistic lock conflict (template modified by another operator)
    delete:
      summary: Delete advertising block (cascades to playlist_items; does NOT delete media; zero cashier tasks)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204':
          description: Advertising block deleted
  /advertising-blocks/{id}/duplicate:
    post:
      summary: Duplicate advertising template in 1 click (clones playlist, sets name to '<Name> — копия')
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '201':
          description: Duplicated advertising template created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdvertisingBlock'
  /advertising-blocks/{id}/preview:
    get:
      summary: Get resolved preview payload for Web UI player (renders static or slideshow preview)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: string, format: uuid }
                  area: { type: string }
                  display_mode: { type: string }
                  items:
                    type: array
                    items:
                      type: object
                      properties:
                        url: { type: string }
                        duration_seconds: { type: integer }
                        order_index: { type: integer }
...
components:
  schemas:
...
    MediaAsset:
      type: object
      properties:
        id: { type: string, format: uuid }
        original_name: { type: string }
        stored_name: { type: string }
        sha256: { type: string }
        mime_type: { type: string }
        media_type: { type: string, enum: [IMAGE, VIDEO] }
        file_size_bytes: { type: integer }
        width: { type: integer, nullable: true }
        height: { type: integer, nullable: true }
        version: { type: integer, description: OCC version counter }
        preview_url: { type: string }
    PlaylistItem:
      type: object
      properties:
        id: { type: string, format: uuid }
        media_asset_id: { type: string, format: uuid }
        order_index: { type: integer }
        duration_seconds: { type: integer }
        media_asset: { $ref: '#/components/schemas/MediaAsset' }
    AdvertisingBlock:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        description: { type: string, nullable: true }
        area: { type: string, enum: [FULL_SCREEN, MODE32_PROMO] }
        display_mode: { type: string, enum: [STATIC, SLIDESHOW, VIDEO] }
        is_active: { type: boolean }
        version: { type: integer, description: OCC version counter }
        playlist:
          type: array
          items: { $ref: '#/components/schemas/PlaylistItem' }
...